refactor: dto 구분
